Adding missing file name
Line 186 references the second file you are to make by name but that is the only place where that filename is given. This leads to confusion if you are following the document like a tutorial, when you arrive at line 125-141 you know you are supposed to add 144-172 to a file, but you don't know which. the previous one? Do you make a new one? it's vague. Then when you get to line 186 you either realize what you were supposed to do earlier or you are left wondering where you were supposed to get "add_apt_repo"

So this little change clarifies all that.