Update connect-azure-firewall.md
Added AzureFirewallDNSProxy (log) to line 3 f in the Connect to Azure Firewall section of the document.

Current: Mark the check boxes of the rule types whose logs you want to ingest. We recommend AzureFirewallApplicationRule and AzureFirewallNetworkRule.

New/Updated: Mark the check boxes of the rule types whose logs you want to ingest. We recommend AzureFirewallApplicationRule, AzureFirewallNetworkRule and AzureFirewallDNSProxy.