Mark commands in headings as code to prevent Acrolinx scoring
If we're not going to capitalize these because they're commands that are typically used in lowercase, then we should be able to mark them as code, too. This has the benefit of preventing Acrolinx from scoring them, including their lack capitalization. If we instead capitalized these commands in the headings, the Acrolinx score would be about 95.