Update deployment-vdi-microsoft-defender-antivirus.md
Made changes to the last line of the PowerShell script could error if $vdmpath is not on the C: drive. The command assumes it is the C: as part of the cmd command to change to that drive letter. A better solution is to use the native cd /d command instead. The /d command changes both drive and directory instead of the basic directory only. The command I suggest below will work no matter what drive is being used.