Clarification on "EDR Block Mode"
When polling the state of Defender via Powershell Get-MpComputerStatus cmdlet in recent versions, AMRunningMode will report "EDR Block Mode" when running in Passive Mode with EDR enabled in Block Mode. Information about this mode was mentioned in the description for Passive Mode but the table does not have an entry for this specific mode and the Get-MpComputerStatus description doesn't mention that Defender service is running in this mode, so attempting to add a little extra clarity.