Creating file for Skill-up training resources
Creating a new file for adding the training resources from the Sentinel Ninja training https://techcommunity.microsoft.com/t5/microsoft-sentinel-blog/become-a-microsoft-sentinel-ninja-the-complete-level-400/ba-p/1246310 into the public documentation. 

// it is not finished yet. 

The structure will remain the same, focusing on the 5 parts of the training, mirroring the Ninja training. 

The Skill-up resources sub-section should be placed under the Resources section in the Sentinel documentation. 

@rayne-wiselman