Update preset-security-policies.md
Added the note about "MBIQ" near the relevant bullet point 5. related to user impersonation. Similar to the note about Accepted Domains/Owned domains exists near the relevant bullet point 6. related to domain impersonation. 

We should either keep both the notes or remove both the notes since the other note is added to the section "### Policy settings in preset security policies". IMO, it is ok to keep them in steps 5 & 6 even though it is duplicate because customers will know the relevance as they read 5 & 6.