Corrected a wrong and misleading limitation
After spending months with Microsoft on behalf of an important public sector organization, Microsoft recognized that the documentation was wrong about the limitation of "100,000 devices and 7 Mb files" which is by essence impossible as the limitation for any OMA-URI XML file in Microsoft back-end is 350,000 characters. 
If you require, I can share with you the official answer from Microsoft Premier support escalated to the product owner. 
The error in the documentation led our client to loose a lot of time and money to find that specific answer. 
Feel free to contact me for any improvement on my pull request.