Update use-spam-notifications-to-release-and-report-quarantined-messages.md
@chrisda  we can add something like this: 

Distribution Group. Messages sent to this Group will be expanded to group members and members will receive Quarantine Notifications email.
Unified Group. Messages sent to this Group will not be expanded, but if the setting "send a copy to group members" is enabled, all the members will receive an email copy.