Update overview-endpoint-detection-response.md
Updated note as per the microsoft support ticket. Defender for Endpoint detection is not intended to be an auditing or logging solution that records every operation or activity that happens on a given machine.Our sensor has an internal throttling mechanism, so the high rate of repeat identical events will not flood the logs.