Update manage-tenant-allow-block-list.md
https://github.com/MicrosoftDocs/microsoft-365-docs-pr/pull/16387
https://microsoft-ce-csi.acrolinx.cloud/api/v1/checking/scorecards/80228487-3b4c-43e3-9bcf-57f957c3ff21#CORRECTNESS
Line 35: the the > the
Line 83: by by > by

Editorial change based on Correctness issue:
Line 87: If the message was not blocked, and allow entry for the sender is not created, so it won't on the Spoofed senders tab or the Domains & addresses tab. 

Options:
- Delete "so" 
or 
- Delete "If" and capitalize "the"

I chose to delete "so." 
If the message was not blocked, and allow entry for the sender is not created, it won't on the **Spoofed senders** tab or the **Domains & addresses** tab.