Update configure-global-settings-for-safe-links.md
The Global Settings Menu is in the process of being deprecated. All of the policy configuration settings except Block URL List have been removed from the menu. Today customers unable to add new entries to the Block URL List. Safe Links has migrated all of the existing entries to TABL at this point and customers are being directed to clean-up the existing list entries that could not be migrated automatically. Can we add a designation at the top of this page that this Menu and Feature are being deprecated?