Update configure-your-spam-filter-policies.md
On the Actions page that appears, configure the following settings:

6. Message actions > Remove "High confidence phishing" checkmark against "Move message to Junk Email folder:" action based on following MC237394 and below documentation.
https://learn.microsoft.com/en-us/microsoft-365/security/office-365-security/secure-by-default?view=o365-worldwide
We have deprecated the Move message to Junk Email folder action for a High confidence phishing email verdict in EOP anti-spam policies. Anti-spam policies that use this action for high confidence phishing messages will be converted to Quarantine message. The Redirect message to email address action for high confidence phishing messages is unaffected.