Update eval-defender-investigate-respond-simulate-attack.md
The URL to download the payload points to a redirect which makes the script returns an error. I added the redirected URL in the script so it does work. The indentation was incorrect as well which you couldn't copy and paste the script without making any adjustments to the script.