Update connect-cef-syslog.md to use python3 instead of python
Modern Linux distributions have generally abandoned the use of the python alias in favor of python3. Some distributions do include the "alternatives" command to set the default python alias but this is only done by default in the older distros (e.g. RHEL/CentOS 7) and only for RHEL compatibles. Debian based distros no longer include a default alias for python. 

Changing this to python to python3 in the example command will improve customer experience as they are very likely to run the command on distributions of Linux that will not have the alias and will thus fail.

This also helps improve documentation consistency as in the Azure Monitor documentation the command uses python3.
https://learn.microsoft.com/en-us/azure/sentinel/forward-syslog-monitor-agent?bc=%2Fazure%2Fazure-monitor%2Fbreadcrumb%2Ftoc.json&toc=%2Fazure%2Fazure-monitor%2Ftoc.json#configure-linux-syslog-daemon