BTP solution migration to unlimited public preview
The BTP solution moved from a limited (sign up based) public preview to regular (unlimited public preview).
Removing the need to sign up and to use the private preview flag.
The updated and ungated solution is fully approved and available in content hub and marketplace. You can update the documentation to reflect it without waiting to any additional confirmation.