EDR in Block Mode - Recommended
I've noticed that across various channels, many field engineers are being informed that there is some level of benefit, even if minimal, when EDR block is enabled alongside Defender AV as the primary antivirus.

It would be helpful to define what constitutes a minimal benefit. If there is indeed a benefit, it's advisable to keep it enabled. However, if there are any drawbacks, please include a note for clarification.