Update configure-process-opened-file-exclusions-microsoft-defender-antivirus.md
Rewrite the section on wildcards to be accurate to current implementation. Describe the difference between image and path exclusions. Add a mention of Contextual Exclusions. Move the wildcard section up to the top alongside the general process exclusion info so we don't have to read through information about Powershell APIs etc. 

Also increase the heading depth by 1 on everything to try and make the structure a bit easier to read.