Update outbound-spam-policies-configure.md
Removing inaccurate verbiage from lines 117 and 120. "There's no way for the admin to override this block.". Due to a change made last semester this is no longer true, Tenant Admins are now able to remove a user from the Restricted Entities page when they have been blocked for a day.