Update comprehensive-guidance-on-linux-deployment.md
Added a link that explains in great detail what the enterprise certificate pinning process Defender for Endpoint uses to keep the traffic between Endpoints and the backend cloud service secure. This is part of the important note that mentions this traffic should not be SSL inspected because it will break the certificate pinning chain.