Update microsoft-secure-score.md
The linked community doesn't exist anymore. Or maybe it does, but not there - navigating Microsoft's community structure is a skill above my pay grade. Linked to the XDR community forums (community hub? community? who knows, really)

Not sure if I missed it, but if not, this article could use a link to wherever bugs/bad Recommendation dialogue can be submitted for fix. AFAIK, it cannot be contributed to by users.