Update safe-attachments-policies-configure.md
Adding a note that leaving user, group and domain field blank applies the policy to all users. The documentation was missing clarity on how to create policy that could apply to all users in the org. A customer with large user and hunderds of domain reported this scenario and was a blocker due to missing clarity here.