Update quarantine-policies.md
There was a customers' confusion we're seeing is specifically around the built-in AdminOnlyAccessPolicy, which is not editable and has the notification setting hardcoded to “Off”. This means:
Admins cannot enable or disable notifications for this policy.
End users should never receive notifications if this default policy is applied.
The problem arises because the documentation presents notification settings as configurable, which leads some customers to believe this includes the built-in AdminOnlyAccessPolicy. As a result, they open cases wondering why end users receive notifications or asking how to turn them off—when in fact, the default policy doesn’t even allow that. (knowing that by analysis, we can share why users are receiving notifications, based on what is set for this kind of threat SCL level, maybe because it is identified SPAM and he doesn't have the AdminOnly action set for SPAM mails in his custom quarantine policy )
 
To reduce this confusion, I suggest we consider explicitly stating in documentation that:
The default AdminOnlyAccessPolicy is immutable
It always has notifications disabled
If notifications are needed alongside restricted access, admins should use a custom policy with "No access" and configure notifications accordingly.

Discussed with our Beta engineer, Mithun and confirmed to update the documentation with the pulled information