(AzureCXP) fixes MicrosoftDocs/azure-docs#435838
Updated line number 137.

From: When users try to access a blocked application, they receive a message informing them that the application was by their organization. This message is customizable.

To: When users try to access a blocked application, they receive a message informing them that the application was blocked by their organization. This message is customizable.