Update validate-antimalware.md
The documentation causes confusion with customers with regard to the detection behavior of Windows Defender as they expect an immediate detection once the file is saved. But due to perf optimizations that make this type of validation quite problematic.

Added instruction line "3. Open a Command Prompt and Run: 'type EICAR.txt' " to make the validation deterministic and reduce confusion of Cx (less IcM).